// dspace-angular/src/app/shared/form/builder/parsers/parser-type.ts
export enum ParserType {
  Date = 'date',
  Dropdown = 'dropdown',
  RelationGroup = 'group',
  List = 'list',
  Lookup = 'lookup',
  LookupName = 'lookup-name',
  LookupRelation = 'lookup-relation',
  Onebox = 'onebox',
  Name = 'name',
  Series = 'series',
  Tag = 'tag',
  Textarea = 'textarea'
}